a start at network API, some refactoring
